+++ enterprise/admin/__init__.py
@@ -211,6 +211,7 @@ class EnterpriseCustomerAdmin(DjangoObjectActions, SimpleHistoryAdmin):
                        'enable_integrated_customer_learner_portal_search',
                        'enable_analytics_screen', 'enable_audit_enrollment',
                        'enable_audit_data_reporting', 'enable_learner_portal_offers',
+                       'disable_expiry_messaging_for_learner_credit',
                        'enable_executive_education_2U_fulfillment',
                        'enable_career_engagement_network_on_learner_portal',
                        'career_engagement_network_message', 'enable_pathways', 'enable_programs',